myproject/requirements.txt
django.shortcuts
django.db
django.shortcuts
django.db
django.core.paginator
django.contrib.auth.decorators
django.shortcuts
channels